Add files via upload
This folder contains the essential python scripts and configurations for implementing a Label and Text Detection System on the AWS cloud platform. From setting up foundational resources with CloudFormation to deploying Lambda functions for image analysis, each file serves a crucial role in the seamless operation of the system. Explore the scripts to gain insights into the intricacies of cloud application development and AWS service integration